requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
pyyaml>=6.0
